+++ xlmm/src/main/java/com/jimei/xiaolumeimei/entities/CartsinfoBean.java
@@ -33,8 +33,8 @@ public class CartsinfoBean {
   @SerializedName("buyer_nick") private String mBuyerNick;
   @SerializedName("num") private String mNum;
   @SerializedName("remain_time") private String mRemaStringime;
-  @SerializedName("std_sale_price") private float mStdSalePrice;
-  @SerializedName("total_fee") private float mTotalFee;
+  @SerializedName("std_sale_price") private double mStdSalePrice;
+  @SerializedName("total_fee") private double mTotalFee;
   @SerializedName("item_id") private String mItemId;
   @SerializedName("pic_path") private String mPicPath;
   @SerializedName("sku_name") private String mSkuName;
@@ -109,7 +109,7 @@ public class CartsinfoBean {
     return mStdSalePrice;
   }
 
-  public void setStdSalePrice(float stdSalePrice) {
+  public void setStdSalePrice(double stdSalePrice) {
     this.mStdSalePrice = stdSalePrice;
   }
 
@@ -117,7 +117,7 @@ public class CartsinfoBean {
     return mTotalFee;
   }
 
-  public void setTotalFee(float totalFee) {
+  public void setTotalFee(double totalFee) {
     this.mTotalFee = totalFee;
   }
